feat: integration tests + sample project for LuaBridge
- Add BreadLua.Sample console project demonstrating LuaState, Buffer<T>, and BindToLua
- Add 6 integration tests (BridgeIntegrationTests) covering DoString, Buffer read/write, BindToLua globals, Call, and error handling
- Fix Buffer.BindToLua overload resolution bug: int _count was matching SetGlobal(string, IntPtr) instead of SetGlobal(string, long) due to .NET nint implicit conversion
- All 16 tests pass

+++ src/BreadLua.Runtime/Core/Buffer.cs
@@ -39,7 +39,7 @@ public unsafe class Buffer<T> : IDisposable where T : unmanaged
     public void BindToLua(LuaState state, string globalName)
     {
         state.SetGlobal(globalName, (IntPtr)_ptr);
-        state.SetGlobal(globalName + "_count", _count);
+        state.SetGlobal(globalName + "_count", (long)_count);
     }
 
     public void Dispose()
